% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa_needed.R
\name{gpa_needed}
\alias{gpa_needed}
\title{GPA Needed Function}
\usage{
gpa_needed(current_gpa, gpa_wanted, terms, plusminus = T)
}
\arguments{
\item{current_gpa}{The student's current cumulative GPA}

\item{gpa_wanted}{The cumulative GPA goal for the end of the term}

\item{terms}{The number of terms the student has been in college. This includes the current term}

\item{plusminus}{Whether or not the grading scale is the normal plus/minus scale}
}
\description{
This function determines the minimum term GPA a student needs to get in order to ensure that their cumulative GPA meets
some goal at the end of the term. If the cumulative GPA goal is unreachable in one term, the function returns a message
to that effect
}
\examples{

}
